hide all first views in codespaces
